Update tier priority to uint8

+++ pkg/apis/security/v1alpha1/types.go
@@ -216,7 +216,7 @@ type Tier struct {
 type TierSpec struct {
 	// Priority specfies the order of the Tier relative to other Tiers.
 	// Currently supporting only 5 Tiers. i.e. 1-5 range for Priority.
-	Priority int32 `json:"priority"`
+	Priority uint8 `json:"priority"`
 	// Description is an optional field to add more information regarding
 	// the purpose of this Tier.
 	Description string `json:"description,omitempty"`
